Fix CurrentURIMetaData
Before, the whole directory has been served instead of just the metadata of the CurrentURI. This worked, because the audio format was the same, so the renderer could interpret it correctly. With the support of multiple formats, it broke if the default format wasn't the same as the first entry's format in the directory list because the renderer interpreted is as the metadata-specified format, not as the format specified in the http headers. By sending empty metadata, the renderer is forced to always read the format from the http headers.