Fix volume button icon switching
This fixes an old bug where the volume button icon wouldn't switch back to unmuted after a mute, resulting from the fact that the icon index was initialized with zero (mute icon) and never updated for the actual volume. This happened because the handler updating the icon index wasn't called if the initial volume equaled the value of the volume slider.